Split csv mutations
