+++ saleor/graphql/csv/tests/mutations/test_export_gift_cards.py
@@ -55,7 +55,7 @@ EXPORT_GIFT_CARDS_MUTATION = """
         ),
     ],
 )
-@patch("saleor.graphql.csv.mutations.export_gift_cards_task.delay")
+@patch("saleor.graphql.csv.mutations.export_gift_cards.export_gift_cards_task.delay")
 def test_export_gift_cards_mutation(
     export_gift_cards_mock,
     input,
@@ -93,7 +93,7 @@ def test_export_gift_cards_mutation(
     ).exists()
 
 
-@patch("saleor.graphql.csv.mutations.export_gift_cards_task.delay")
+@patch("saleor.graphql.csv.mutations.export_gift_cards.export_gift_cards_task.delay")
 def test_export_gift_cards_mutation_ids_scope(
     export_gift_cards_mock,
     staff_api_client,
@@ -151,7 +151,7 @@ def test_export_gift_cards_mutation_ids_scope(
     ).exists()
 
 
-@patch("saleor.graphql.csv.mutations.export_gift_cards_task.delay")
+@patch("saleor.graphql.csv.mutations.export_gift_cards.export_gift_cards_task.delay")
 def test_export_gift_cards_mutation_ids_scope_invalid_object_type(
     export_gift_cards_mock,
     staff_api_client,
@@ -221,7 +221,7 @@ def test_export_gift_cards_mutation_ids_scope_invalid_object_type(
         ),
     ],
 )
-@patch("saleor.graphql.csv.mutations.export_gift_cards_task.delay")
+@patch("saleor.graphql.csv.mutations.export_gift_cards.export_gift_cards_task.delay")
 def test_export_gift_cards_mutation_failed(
     export_gift_cards_mock,
     input,
@@ -286,7 +286,7 @@ EXPORT_GIFT_CARDS_MUTATION_BY_APP = """
 """
 
 
-@patch("saleor.graphql.csv.mutations.export_gift_cards_task.delay")
+@patch("saleor.graphql.csv.mutations.export_gift_cards.export_gift_cards_task.delay")
 def test_export_gift_cards_mutation_by_app(
     export_gift_cards_mock,
     app_api_client,
